+++ src/components/githubPage/FeatureComparisonSection/FeatureComparisonSection.module.scss
