+++ src/components/githubPage/FeatureComparisonSection/FeatureComparisonSection.tsx
